+++ test/vhpi/vhpi_test.h
@@ -33,5 +33,6 @@ void vhpi1_startup(void);
 void vhpi2_startup(void);
 void vhpi3_startup(void);
 void vhpi5_startup(void);
+void vhpi6_startup(void);
 
 #endif  // _VHPI_TEST_H