+++ NEWS.md
@@ -15,6 +15,8 @@
 - The [xpm_vhdl](https://github.com/fransschreuder/xpm_vhdl) project
   which provides VHDL models of the Xilinx XPM macros can now be
   installed with `nvc --install xpm_vhdl`.
+- Many improvements to the VHPI implementation (from @Forty-Bot).
+- `vhpi_put_value` with `vhpiDepositPropagate` mode is now supported.
 
 ## Version 1.9.2 - 2023-05-01
 - Fix elaboration errors with recursive entity instantiation (#668).